requirements.txt
# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9

# Configuration
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1

# Email
fastapi-mail==1.4.1

# File Storage (S3)
boto3==1.34.101

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0

# Utilities
python-dateutil==2.9.0
